+++ HISTORY.md
@@ -1,9 +1,15 @@
 0.1 Release
 -----------
 
-* Changed ``Record`` to use an ordered dict of samples called ``genotypes``.
 * Added test code
-
+* Added Writer class
+* Allow negative number in ``INFO`` and ``FORMAT`` fields (thanks @martijnvermaat)
+* Prefer ``vcf.Reader`` to ``vcf.VCFReader``
+* Support compressed files with guessing where filename is available on fsock
+* Allow opening by filename as well as filesocket
+* Support fetching rows for tabixed indexed files
+* Performance improvements (see ``test/prof.py``)
+* Added extensible filter script (see FILTERS.md), vcf_filter.py 
 
 Contributions
 -------------